+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Version History
 
-## 2.0.0
+## Development
+
+* The ConfigDB watcher now watches DataProductPersist Flows and checks the status of the related DataProduct Flow.
 
 ### Updated